Fix positional parameters in queries.
Placeholder for parameters must be used in the form of `?n` where n is
an integer. Instead of checking for the total number of placeholders,
the max `n` must be selected as number of expected parameters. This
fixes the issue where if you only use `?1` three times in your SQL
string, then an error is raised that 3 parameters are expected.